==> code/4_filter_rnaseq_vdj_data.Rmd <==
---
title: "Filter scRNA-Seq and VDJ data"
output: html_document
date: "2024-07-02"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# START
* Workflow for processing combined analysis object
* Prior to running this code set working directory to 'impact-analysis/code'

```{r,warning=F,message=F}
# Set Working Directory before starting

# Immcantation
suppressPackageStartupMessages(library(alakazam))
suppressPackageStartupMessages(library(shazam))
suppressPackageStartupMessages(library(tigger))
suppressPackageStartupMessages(library(dowser))
suppressPackageStartupMessages(library(airr))

suppressPackageStartupMessages(library(rjson))
suppressPackageStartupMessages(library(glmGamPoi))
suppressPackageStartupMessages(library(Seurat))
suppressPackageStartupMessages(library(dittoSeq))
suppressPackageStartupMessages(library(reshape2))
suppressPackageStartupMessages(library(kableExtra))
suppressPackageStartupMessages(library(devtools))
suppressPackageStartupMessages(library(data.table))
suppressPackageStartupMessages(library(ggplot2))
suppressPackageStartupMessages(library(RColorBrewer))
suppressPackageStartupMessages(library(gridExtra))
suppressPackageStartupMessages(library(cowplot))
suppressPackageStartupMessages(library(dplyr))
suppressPackageStartupMessages(library(impactSingleCellToolkit))

# Increase size of ENV
options(future.globals.maxSize= 891289600)
set.seed(12345)
```


# 1. Load inputs

### a. params
```{r}
# INPUT and OUTPUT Directories
param_file_fh = "../input/input_one_patient_analysis.json"
params        = fromJSON(file = param_file_fh)

# INPUT
metadata_fh   = params$metadata_file
metadata      = read.csv(metadata_fh,stringsAsFactors = F,check.names = F)
THREADS       = params$threads

# subset metadata
metadata_gex  <- metadata[metadata$type %in% c("counts_gex"),]
metadata_bcr  <- metadata[metadata$type %in% c("vdj_b"),]
metadata_tcr  <- metadata[metadata$type %in% c("vdj_t"),]

# OUTPUT Results Directories
plot_dir     = params$plot_directory
objects_dir  = params$objects_directory
if(!file.exists(plot_dir)){dir.create(plot_dir,recursive = TRUE)}
if(!file.exists(objects_dir)){dir.create(objects_dir,recursive = TRUE)}
```

### b. processed GEX/CSP object
```{r}
fh_processed_seurat_obj <- file.path(objects_dir,"seurat_obj.processed.rds")
full_rnaseq_obj         <- readRDS(fh_processed_seurat_obj)
```


# 2. Load and format TCR/BCR repertoire results
* Clonotype definitions for BCR are based on the Heavy chain only
  * Plots CDR3 distance distribution among BCRs and come up with a new threshold
* Clonotype definitions for TCR include both Alpha and Beta chains

### a. Load Immcantation Results
```{r}
# File paths
bcr_heavy_fh <- "../../Results/results_immcantation_full_cohort/BCR_CSFPB/BCR_CSFPB_heavy_germ-pass.tsv"
bcr_light_fh <- "../../Results/results_immcantation_full_cohort/BCR_CSFPB/BCR_CSFPB_light_germ-pass.tsv"
tcr_beta_fh  <- "../../Results/results_immcantation_one_patient/TCR_CSFPB/TCR_CSFPB_heavy_germ-pass.tsv"
tcr_alpha_fh <- "../../Results/results_immcantation_one_patient/TCR_CSFPB/TCR_CSFPB_light_germ-pass.tsv"

# TCR
db.tra   <- read_rearrangement(tcr_alpha_fh) %>% as.data.frame()
db.trb   <- read_rearrangement(tcr_beta_fh) %>% as.data.frame()
# BCR
db.heavy <- read_rearrangement(bcr_heavy_fh) %>% as.data.frame()
db.light <- read_rearrangement(bcr_light_fh) %>% as.data.frame()
```

### b. filter contigs and merge all chains
* All contigs are in frame
```{r}
bcr_df <- mergeVDJresults(db.heavy,db.light,umi.thresh = 3,assay = "bcr")
tcr_df <- mergeVDJresults(db.trb,db.tra,umi.thresh = 2, assay = "tcr")

# format VDJ unique cell IDs to match scRNA-Seq (function only for this workshop)
formatUniqueIDcolum <- function(uni_id_column,sep = ":") {
  samples <- tstrsplit(uni_id_column, sep)[[1]]
  cells   <- tstrsplit(uni_id_column, sep)[[2]]
  fmt_ids <- paste(cells,samples,sep = "-")
  
  return(fmt_ids)
}

bcr_df$unique_cell_id <- formatUniqueIDcolum(bcr_df$unique_cell_id)
tcr_df$unique_cell_id <- formatUniqueIDcolum(tcr_df$unique_cell_id)
```


# 3. Only analyze cells shared across all datasets
* Separate remaining cell types into another object for further analysis

### a. get vector of overlapping cells
```{r}
cells_rnaseq      <- colnames(full_rnaseq_obj)
cells_vdj         <- unique(c(bcr_df$unique_cell_id,
                              tcr_df$unique_cell_id))
overlapping_cells <- cells_rnaseq[cells_rnaseq %in% cells_vdj]
```


### b. filter rna-seq and vdj data
* Even after filtering for cells that overlap with BCR/TCR data, there are still sparse clusters of cells annotated as different cell types
```{r,fig.height=6,fig.width=14}
# generate UMAP of cell type annotations pre-filtering
p1_pre_filter <- DimPlot(object = full_rnaseq_obj, group.by="predicted.celltype.l1", pt.size=1.0, label.size = 5, alpha = 0.7,label = T,reduction = "umap")

# only keep cells with vdj
rnaseq_obj       <- full_rnaseq_obj[,colnames(full_rnaseq_obj) %in% overlapping_cells]
bcr_df           <- bcr_df[bcr_df$unique_cell_id %in% overlapping_cells,]
tcr_df           <- tcr_df[tcr_df$unique_cell_id %in% overlapping_cells,]

# post filter umap
p2_post_filter <- DimPlot(object = rnaseq_obj, group.by="predicted.celltype.l1", pt.size=1.0, label.size = 5, alpha = 0.7,label = T,reduction = "umap")

plot_grid(p1_pre_filter,p2_post_filter,ncol = 2)
```


# 4. Add metadata columns to GEX and VDJ data

### a. rnaseq object
```{r}
createMetadataColumn <- function(sample_vec, metadata,sample.col = "sample",meta.col = "SampleType") {
  new_col <- sample_vec
  uni_labels <- metadata[,meta.col] %>% unique()
  
  for (categ in uni_labels) {
    categ_samples <- metadata[metadata[,meta.col] %in% categ,sample.col] %>% unique()
    new_col[new_col %in% categ_samples] <- categ
  }
  
  return(new_col)
}

# Full object
full_rnaseq_obj@meta.data$compartment <- createMetadataColumn(full_rnaseq_obj@meta.data$sample,metadata = metadata_gex,meta.col = "SampleType" )
full_rnaseq_obj@meta.data$status <- createMetadataColumn(full_rnaseq_obj@meta.data$sample,metadata = metadata_gex,meta.col = "Project" )

# VDJ Overlapped object
rnaseq_obj@meta.data$compartment <- createMetadataColumn(rnaseq_obj@meta.data$sample,metadata = metadata_gex,meta.col = "SampleType" )
rnaseq_obj@meta.data$status <- createMetadataColumn(rnaseq_obj@meta.data$sample,metadata = metadata_gex,meta.col = "Project" )
```

### b. VDJ
```{r}
# Add sample column
tcr_df$sample <- tstrsplit(tcr_df$unique_cell_id,"-")[[2]]
bcr_df$sample <- tstrsplit(bcr_df$unique_cell_id,"-")[[2]]

# Add column for Compartment
tcr_df$COMPARTMENT <- createMetadataColumn(tcr_df$sample,metadata = metadata_tcr,meta.col = "SampleType" )
bcr_df$COMPARTMENT <- createMetadataColumn(bcr_df$sample,metadata = metadata_bcr,meta.col = "SampleType" )

# Sample status
tcr_df$STATUS <- createMetadataColumn(tcr_df$sample,metadata = metadata_tcr,meta.col = "Project" )
bcr_df$STATUS <- createMetadataColumn(bcr_df$sample,metadata = metadata_bcr,meta.col = "Project" )
```


# 5. Re-cluster

### a. SCT
```{r}
# Re-cluster
opt_pcs    <- plotOptimalPCsforSeuratObject(rnaseq_obj)
rnaseq_obj <- scaleAndClusterSeuratObject(rnaseq_obj,dims = 1:11,npca = 11,resolution = 0.8,
                                          normalize = F,dim.reduction = T)

# Plot
reclus_umap1 <- dittoDimPlot(rnaseq_obj, assay="RNA", reduction="umap", var="predicted.celltype.l1")
reclus_umap2 <- dittoDimPlot(rnaseq_obj, assay="RNA", reduction="umap", var="seurat_clusters")

plot_grid(reclus_umap1,reclus_umap2,ncol = 2)
```

### b. CSP
```{r}
# Re-cluster
DefaultAssay(rnaseq_obj) <- "CSP" # it should be 'SCT' after processing the GEX data

### Run PCA
VariableFeatures(rnaseq_obj) = rownames(rnaseq_obj@assays[["CSP"]])
rnaseq_obj = rnaseq_obj %>% 
  RunPCA(reduction.name="apca",approx=F, verbose=F) 
### Run UMAP
rnaseq_obj <- RunUMAP(rnaseq_obj, 
              reduction = 'apca', 
              dims = 1:6, 
              assay = 'CSP', 
              reduction.name = 'csp.umap', 
              reduction.key = 'cspUMAP_',
              verbose = F)

DefaultAssay(rnaseq_obj) <- "SCT"


# Plot 
reclus_csp_umap1 <- dittoDimPlot(rnaseq_obj, assay="CSP", reduction="csp.umap", var="predicted.celltype.l1")
reclus_csp_umap2 <- dittoDimPlot(rnaseq_obj, assay="CSP", reduction="csp.umap", var="seurat_clusters")


plot_grid(reclus_csp_umap1,reclus_csp_umap2,ncol = 2)
```

# 6. WNN to get multi-omics clusters
```{r}
#DefaultAssay(rnaseq_obj)

rnaseq_obj <- FindMultiModalNeighbors(
  rnaseq_obj, 
  reduction.list = list("pca", "apca"), # RNA and CSP
  dims.list = list(1:11, 1:6),
  k.nn = 20,
  prune.SNN = 1/15 
)

rnaseq_obj <- RunUMAP(rnaseq_obj, 
              nn.name = "weighted.nn", 
              reduction.name = "wnn.umap", 
              reduction.key = "wnnUMAP_",
              verbose=F)
rnaseq_obj <- FindClusters(rnaseq_obj, 
                   graph.name = "wsnn", 
                   algorithm = 3,
                   resolution = 2, 
                   verbose = F, 
                   random.seed = 12345)

rnaseq_obj@meta.data$wsnn_cluster = rnaseq_obj@meta.data$wsnn_res.2
```


# 7. Explore differences in the clustering

### a. clusters - plot all 3 UMAPs
```{r}
p1 = dittoDimPlot(rnaseq_obj, var="wsnn_cluster", reduction="wnn.umap")+NoLegend()
p2 = dittoDimPlot(rnaseq_obj, var="wsnn_cluster", reduction="csp.umap")+NoLegend()
p3 = dittoDimPlot(rnaseq_obj, var="wsnn_cluster", reduction="umap")+NoLegend()+xlab("rnaUMAP_1")+ylab("rnaUMAP_2")

plot_grid(p1, p2, p3, ncol=3)
```

### b. celltype - plot all 3 UMAPs
```{r,fig.height=5, fig.width=15}
c1 = dittoDimPlot(rnaseq_obj, var="predicted.celltype.l1", reduction="wnn.umap")
c2 = dittoDimPlot(rnaseq_obj, var="predicted.celltype.l1", reduction="csp.umap")
c3 = dittoDimPlot(rnaseq_obj, var="predicted.celltype.l1", reduction="umap")+xlab("rnaUMAP_1")+ylab("rnaUMAP_2")

plot_grid(c1, c2, c3, ncol=3)
```

# 8. Separate B & T cells using VDJ
```{r}
# Add column separating BCR and TCR
vdj_status <- row.names(rnaseq_obj@meta.data)
vdj_status[vdj_status %in% unique(bcr_df$unique_cell_id)] <- "BCR"
vdj_status[vdj_status %in% unique(tcr_df$unique_cell_id)] <- "TCR"
rnaseq_obj$vdj_status <- vdj_status

bcell_cell_ids <- row.names(rnaseq_obj@meta.data[rnaseq_obj@meta.data$vdj_status %in% "BCR",])
tcell_cell_ids <- row.names(rnaseq_obj@meta.data[rnaseq_obj@meta.data$vdj_status %in% "TCR",])

# separate B and T cells for RNA-Seq
bcell_obj     <- rnaseq_obj[,colnames(rnaseq_obj) %in% bcell_cell_ids]
tcell_obj     <- rnaseq_obj[,colnames(rnaseq_obj) %in% tcell_cell_ids]
```


# 9. Plot B/T cells

### a. B cells
```{r,fig.height=5, fig.width=15}
b1 = dittoDimPlot(bcell_obj, var="predicted.celltype.l1", reduction="wnn.umap")
b2 = dittoDimPlot(bcell_obj, var="predicted.celltype.l1", reduction="csp.umap")
b3 = dittoDimPlot(bcell_obj, var="predicted.celltype.l1", reduction="umap")+xlab("rnaUMAP_1")+ylab("rnaUMAP_2")

plot_grid(b1, b2, b3, ncol=3)
```

### b. T cells
```{r,fig.height=5, fig.width=15}
t1 = dittoDimPlot(tcell_obj, var="predicted.celltype.l1", reduction="wnn.umap")
t2 = dittoDimPlot(tcell_obj, var="predicted.celltype.l1", reduction="csp.umap")
t3 = dittoDimPlot(tcell_obj, var="predicted.celltype.l1", reduction="umap")+xlab("rnaUMAP_1")+ylab("rnaUMAP_2")

plot_grid(t1, t2, t3, ncol=3)
```

# 10. Assess hamming distances among BCR data
* These results suggest we should run the full cohort through immcantation and subset the results as needed
* This would make the automated threshold for B cells more reliable, otherwise the distribution is too sparse

### a. generate plots
```{r}
bcr_results_igh <- bcr_df[bcr_df$locus %in% "IGH",]
dist_nearest    <- distToNearest(bcr_results_igh)

# Automated Threshold - will take a long time with the full cohort
# - Threshold setting occurs by default when running the immcantation pipeline
threshold_output <- shazam::findThreshold(dist_nearest$dist_nearest,
                                  method = "gmm", model = "gamma-norm",
                                  cutoff = "user",spc = 0.99)
threshold <- threshold_output@threshold

p_ham_auto <- plot(threshold_output, binwidth = 0.02, silent = TRUE) + theme(axis.title = element_text(size = 18))

# generate Hamming distance histogram
p_ham <- ggplot2::ggplot(subset(dist_nearest, !is.na(dist_nearest)),
             aes(x = dist_nearest)) +
        geom_histogram(color = "white", binwidth = 0.02) +
        labs(x = "Hamming distance", y = "Count") +
        scale_x_continuous(breaks = seq(0, 1, 0.1)) +
        theme_bw() +
        theme(axis.title = element_text(size = 18)) + geom_vline(xintercept=0.15, linetype = "longdash",color = "red")

plot(p_ham)
plot(p_ham_auto)
```

### b. PDF
```{r}
ham_plot_file_name <- paste("distn_hamming_CDR3_auto_thresh_",threshold,".pdf",sep = "")
pdf(file = file.path(plot_dir,ham_plot_file_name),height = 5,width = 7)
plot(p_ham_auto)
dev.off()
```


# Save Objects
```{r}
# Full processed object all cell types

# B and T cell Objects
saveRDS(bcell_obj,file = file.path(objects_dir,"bcell_seurat_obj.rds"))
saveRDS(tcell_obj,file = file.path(objects_dir,"tcell_seurat_obj.rds"))

# Filtered VDJ results
save(bcr_df,tcr_df,threshold_output,file = file.path(objects_dir,"vdj_results.processed.RData"))
```
